backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
google-genai
python-dotenv==1.0.1

